FIX: correct types

+++ entity_linking/data/mention_entity_matching.py
@@ -17,9 +17,9 @@ class MentionEntityMatchingDataset(Dataset):
     def __getitem__(self, idx):
         item = {key: torch.tensor(val[idx]) for key, val in self.encodings.items()}
         item.update({
-            'mention_spans': torch.tensor(self.mention_spans[idx]),
-            'entity_indices': torch.tensor(self.entity_indices[idx]),
-            'labels': torch.tensor(self.entity_labels[idx])
+            'mention_spans': torch.tensor(self.mention_spans[idx], dtype=torch.int),
+            'entity_indices': torch.tensor(self.entity_indices[idx], dtype=torch.int),
+            'labels': torch.tensor(self.entity_labels[idx], dtype=torch.float)
         })
         return item
 
@@ -85,7 +85,7 @@ def _process_entity_info(entity_info: List[List[Tuple[int, Tuple[int, int], list
     num_chunks = len(entity_info)
     mention_spans = np.zeros((num_chunks, num_ents, 2), dtype=int)
     entity_indices = np.zeros((num_chunks, num_ents), dtype=int)
-    entity_labels = np.ones((num_chunks, num_ents), dtype=int) * -100  # initialize all with label: ignore
+    entity_labels = np.ones((num_chunks, num_ents)) * -100  # initialize all with label: ignore
     ents_per_item = num_ents // items_per_chunk
     # prepare word blocker to retrieve entities with similar surface forms
     dbr = DbpResourceStore.instance()
